Added loading skeleton to edit to prevent async bugs
